refactor: HandlerMapping 위치 변경
