+++ mvc/src/main/java/com/interface21/webmvc/servlet/HandlerMapping.java
@@ -1,9 +1,8 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 
 public interface HandlerMapping {
     Object getHandler(HttpServletRequest request);
-
     void initialize();
 }
